- Remove castVote - Remove maxVoterWeightRecord
