+++ programs/gateway/src/instructions/mod.rs
@@ -4,11 +4,5 @@ mod create_registrar;
 pub use create_voter_weight_record::*;
 mod create_voter_weight_record;
 
-pub use create_max_voter_weight_record::*;
-mod create_max_voter_weight_record;
-
 pub use update_voter_weight_record::*;
 mod update_voter_weight_record;
-
-pub use cast_vote::*;
-mod cast_vote;
