+++ programs/gateway/src/state/mod.rs
@@ -1,7 +1,5 @@
 pub use registrar::*;
 pub mod registrar;
 
-pub mod max_voter_weight_record;
-
 pub use voter_weight_record::*;
 pub mod voter_weight_record;
